Refactor MainApplicationFrame separating its logic into different modules
